+++ libs/ui/src/lib/containers/Card/CardFooter.tsx
@@ -6,16 +6,12 @@ export type CardFooterProps = {
   addendum?: boolean;
   action?: React.ReactNode;
   description?: string;
-}
+};
 
 const Card = ({ action, description, addendum = false }: CardFooterProps) => {
   return (
     <S.CardFooter addendum={addendum}>
-      {description && (
-        <S.CardFooterText>
-          {description}
-        </S.CardFooterText>
-      )}
+      {description && <S.CardFooterText>{description}</S.CardFooterText>}
 
       {action && <>{action}</>}
     </S.CardFooter>