+++ libs/ui/src/lib/forms/Action/Action.stories.tsx
@@ -7,8 +7,14 @@ export default {
   component: Action,
 };
 
-export const IconLeading = () => <Action label="Action Label" arrowPosition="leading" />;
+export const IconLeading = () => (
+  <Action label="Action Label" arrowPosition="leading" />
+);
 
-export const IconTrailling = () => <Action label="Action Label" arrowPosition="trailing" />;
+export const IconTrailling = () => (
+  <Action label="Action Label" arrowPosition="trailing" />
+);
 
-export const Wide = () => <Action label="Action Label" arrowPosition="trailing" wide />;
+export const Wide = () => (
+  <Action label="Action Label" arrowPosition="trailing" wide />
+);