+++ libs/ui/src/lib/forms/Action/Action.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { ButtonProps } from "../Button/Button";
+import { ButtonProps } from '../Button/Button';
 
-import ArrowLeft from "./ArrowLeft.icon";
-import ArrowRight from "./ArrowRight.icon";
+import ArrowLeft from './ArrowLeft.icon';
+import ArrowRight from './ArrowRight.icon';
 
 import * as S from './Action.style';
 
@@ -13,7 +13,12 @@ type ActionProps = {
   wide?: boolean;
 } & ButtonProps;
 
-export const Action = ({ label, wide, arrowPosition, ...rest}: ActionProps) => (
+export const Action = ({
+  label,
+  wide,
+  arrowPosition,
+  ...rest
+}: ActionProps) => (
   <S.Action
     label={label}
     arrowPosition={arrowPosition}